Reorganizing files into the following structure: - lib - modules - vendor
