--- Assets/src/js/forms/isValidEmailAddress.js
@@ -1,10 +1,0 @@
-// @isValidEmailAddress
-//
-// example:
-// if(isValidEmailAddress(theEnteredEmailAddy) != true) { tell em there's an error }
-// returns true if email is valid.
-//
-function isValidEmailAddress(emailAddress) {
-	var pattern = new RegExp(/^(("[\w-\s]+")|([\w-]+(?:\.[\w-]+)*)|("[\w-\s]+")([\w-]+(?:\.[\w-]+)*))(@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$)|(@\[?((25[0-5]\.|2[0-4][0-9]\.|1[0-9]{2}\.|[0-9]{1,2}\.))((25[0-5]|2[0-4][0-9]|1[0-9]{2}|[0-9]{1,2})\.){2}(25[0-5]|2[0-4][0-9]|1[0-9]{2}|[0-9]{1,2})\]?$)/i);
-	return pattern.test(emailAddress);
-}
