+++ Assets/src/js/head.js
@@ -1,1 +1,1 @@
-require('./modernizr');
+require('./vendor/modernizr');
