+++ Assets/src/js/lib/README.md
@@ -1,0 +1,2 @@
+# Lib Directory
+The lib directory is where custom, common libraries will exist.  Examples could include utilities that get, set, and delete cookies.
