+++ Assets/src/js/lib/alternativeDocReady.js
