+++ Assets/src/js/lib/breakingNonSpace.js
