+++ Assets/src/js/lib/changeText.js
