+++ Assets/src/js/lib/charactersRemaining.js
