+++ Assets/src/js/lib/flickerPrevention.js
