+++ Assets/src/js/lib/foriegnLinks.js
