+++ Assets/src/js/lib/getScript.js
