+++ Assets/src/js/lib/inputClear.js
