+++ Assets/src/js/lib/insertAfter.js
