+++ Assets/src/js/lib/minHeight.js
