+++ Assets/src/js/lib/scoundrels.js
