+++ Assets/src/js/lib/suffler.js
