+++ Assets/src/js/modules/README.md
@@ -1,0 +1,2 @@
+# Modules Directory
+The modules directory is for custom developed modules.  These may includes items such as wrappers and extensions for jQuery plugins or entirely new classes that you've built.
